Fixed the filter, cleaned up file structure slightly, committing so I don't lose the filter again.
