+++ src/components/dashboard/Filter.js
@@ -1,46 +1,15 @@
 import React, {useState} from 'react';
-import 'antd/dist/antd.css'
-import {Menu, Button, Row, Col} from 'antd';
+import { Menu, Button } from 'antd';
 import styled from 'styled-components';
-import {
-  MenuUnfoldOutlined,
-  MenuFoldOutlined,
-} from '@ant-design/icons';
 
-const { SubMenu } = Menu;
-
-
-export const Heading = () =>{
-
-  const [collapsed, setCollapsed] = useState([false]);
-  function toggleCollapsed(){
-    setCollapsed(!collapsed)
-  }
-
-  return (
-    <div>
-      <Row>
-          <Col span={20}>
-              <StyledSpan>Overview of biodiversity in the Congo Basin Rainforest</StyledSpan>
-          </Col>
-          <Col span={4}>
-              <StyledButton onClick={toggleCollapsed}><span>Filter</span></StyledButton>
-          </Col>  
-      </Row>
-      <Filter collapsed={collapsed} toggleCollapsed={toggleCollapsed}/>
-    </div>
-
-  )
-}
-
-function Filter(props){ 
+export default function Filter(props){ 
   return (
         <Menu
           mode="inline"
           inlineCollapsed={props.collapsed}
           selectable='true'
           multiple='true'
-          style={{position:'fixed', float:'right', top:0, zIndex:1 }}
+          style={{width:'20vw', height:'100vh'}}
         >
           <Menu.ItemGroup
             key="sub1"
@@ -60,7 +29,7 @@ function Filter(props){
 
 const StyledButton = styled(Button)`
   background-color: #45735D;
-  width:100%;
+  width:50%;
   border-radius:5px;
   padding:0;
   margin:0 auto;
@@ -72,10 +41,4 @@ const StyledButton = styled(Button)`
     vertical-align:middle;
   }
   
-`;
-
-const StyledSpan = styled.span`
-  font-size:2em;
-  font-weight:bold;
-  color:black;
 `;
